+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Version format is `major`.`minor`.`patch`. Until major version 1 is realeased, A
 
 ## 0.10.0
 - Add `deleting` state to meetings
-- Allow filtering on `deleting`, `archiving/archived` and `closed` meetings in Find meeting
+- Allow filtering on meeting states in Find meeting [#219](https://github.com/VoteIT/voteit_frontend/issues/219)
 - Rename Settings to Control panel
 - Add Export from meeting in Control panel, closed [#218](https://github.com/VoteIT/voteit_frontend/issues/218)